refactor(server-core): inline route handlers into controller methods
Move handler logic from separate handler files directly into controller
class methods for analysis-log, analysis-node-log, and root status.
Removes the handlers/ and status/ directories — all controllers now
follow the same thin-controller pattern with logic in the method body.
