--- apps/server-core/src/adapters/http/controllers/entities/analysis-log/handlers/read.ts
@@ -1,76 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import type { Log, LogLevel, APIClient as TelemetryClient } from '@privateaim/telemetry-kit';
-import { LogFlag } from '@privateaim/telemetry-kit';
-import { useRequestQuery } from '@routup/basic/query';
-import type { IRoutupEvent } from 'routup';
-import { BadRequestError } from '@ebec/http';
-import { type FiltersBuildInput, parseQuery } from 'rapiq';
-import type { AnalysisLog } from '@privateaim/core-kit';
-import { DomainType } from '@privateaim/core-kit';
-
-export async function getManyAnalysisLogRouteHandler(event: IRoutupEvent, telemetryClient?: TelemetryClient) : Promise<any> {
-    const output = parseQuery<AnalysisLog>(
-        useRequestQuery(event),
-        {
-            defaultPath: 'log',
-            filters: {
-                allowed: [
-                    'level',
-                    'analysis_id',
-                ],
-            },
-            pagination: { maxLimit: 50 },
-            sort: { allowed: ['time'] },
-        },
-    );
-
-    // todo: clean up this
-
-    const filtersNormalized : Partial<Record<keyof AnalysisLog, string>> = {};
-    if (output.filters) {
-        for (let i = 0; i < output.filters.length; i++) {
-            filtersNormalized[output.filters[i].key] = `${output.filters[i].value}`;
-        }
-    }
-
-    if (!filtersNormalized.analysis_id) {
-        throw new BadRequestError('The filter analysis_id must be defined.');
-    }
-
-    const filters : FiltersBuildInput<Log> = {
-        labels: {
-            [LogFlag.REF_TYPE]: DomainType.ANALYSIS,
-            [LogFlag.REF_ID]: filtersNormalized.analysis_id,
-        },
-    };
-
-    if (filtersNormalized.level) {
-        filters.level = filtersNormalized.level as LogLevel;
-    }
-
-    // todo: sort missing
-
-    if (telemetryClient) {
-        return telemetryClient.log.getMany({
-            filters,
-            pagination: output.pagination,
-        });
-    }
-
-    return {
-        data: [],
-        meta: {
-            total: 0,
-            pagination: {
-                limit: 50,
-                offset: 0,
-            },
-        },
-    };
-}
