--- apps/server-core/src/adapters/http/controllers/entities/analysis-node-log/handlers/create.ts
@@ -1,96 +1,0 @@
-/*
- * Copyright (c) 2022-2024.
- * Author Peter Placzek (tada5hi)
- * For the full copyright and license information,
- * view the LICENSE file that was distributed with this source code.
- */
-
-import { pickRecord } from '@authup/kit';
-import { DomainType } from '@privateaim/core-kit';
-import { isRealmResourceWritable } from '@privateaim/kit';
-import { ForbiddenError } from '@ebec/http';
-import type { Log, LogLevel, APIClient as TelemetryClient } from '@privateaim/telemetry-kit';
-import { LogChannel, LogFlag } from '@privateaim/telemetry-kit';
-import type { IRoutupEvent } from 'routup';
-import { useDataSource } from 'typeorm-extension';
-import { useRequestIdentityRealm } from '@privateaim/server-http-kit';
-import { readRequestBody } from '@routup/basic/body';
-import { ValidupError, buildErrorMessageForAttribute, defineIssueItem } from 'validup';
-import {
-    AnalysisEntity,
-    NodeEntity,
-} from '../../../../../database/index.ts';
-import { AnalysisNodeLogValidator } from '../utils/index.ts';
-
-export async function createAnalysisNodeLogRouteHandler(event: IRoutupEvent, telemetryClient?: TelemetryClient) : Promise<any> {
-    const validator = new AnalysisNodeLogValidator();
-    const body = await readRequestBody(event);
-    const data = await validator.run(body, { group: 'create' });
-
-    const dataSource = await useDataSource();
-    const nodeRepository = dataSource.getRepository(NodeEntity);
-    const node = await nodeRepository.findOneBy({ id: data.node_id });
-    if (!node) {
-        throw new ValidupError([
-            defineIssueItem({
-                path: ['node_id'],
-                message: buildErrorMessageForAttribute('node_id'),
-            }),
-        ]);
-    }
-
-    data.node_realm_id = node.realm_id;
-
-    const analysisRepository = dataSource.getRepository(AnalysisEntity);
-    const analysis = await analysisRepository.findOneBy({ id: data.analysis_id });
-    if (!analysis) {
-        throw new ValidupError([
-            defineIssueItem({
-                path: ['analysis_id'],
-                message: buildErrorMessageForAttribute('analysis_id'),
-            }),
-        ]);
-    }
-
-    data.analysis_realm_id = analysis.realm_id;
-
-    const isAuthorityOfNode = isRealmResourceWritable(useRequestIdentityRealm(event), data.node_realm_id);
-    if (!isAuthorityOfNode) {
-        throw new ForbiddenError('You are not an actor of to the node realm.');
-    }
-
-    const labels : Record<string, string> = {};
-    const labelsRaw = {
-        ...(data.labels || {}),
-        [LogFlag.REF_TYPE]: DomainType.ANALYSIS_NODE,
-        ...pickRecord(data, [
-            'analysis_id',
-            'node_id',
-            'code',
-            'status',
-        ]),
-    };
-
-    const keys = Object.keys(labelsRaw);
-    for (const key of keys) {
-        if (typeof labelsRaw[key] === 'string') {
-            labels[key] = labelsRaw[key];
-        }
-    }
-
-    const entity : Log = {
-        time: new Date().toISOString(),
-        level: data.level as LogLevel,
-        channel: LogChannel.HTTP,
-        service: 'hub-server-core',
-        message: data.message,
-        labels,
-    };
-
-    if (telemetryClient) {
-        await telemetryClient.log.create(entity);
-    }
-
-    event.response.status = 202;
-    return entity;
-}
